Refactor api.js to use separate routes
